--- src/components/LateNight/LateNightAllow/style.ts
@@ -1,17 +1,0 @@
-import styled from "styled-components";
-
-export const LateNightHeaderContainer = styled.div`
-  display: flex;
-
-  gap: 950px;
-`;
-export const SearchBarContainer = styled.div`
-  display: flex;
-`;
-export const SelectContainer = styled.div``;
-export const InfoText = styled.p`
-  font-size: 12px;
-  color: red;
-
-  margin: 4px 0px 0px 10px;
-`;
