--- src/components/LateNight/LateNightModal/connstant.ts
@@ -1,24 +1,0 @@
-export const LATENIGHT_MODAL_ITEMS = [
-  "이름",
-  "학반",
-  "시작일",
-  "종료일",
-  "장소",
-  "휴대폰",
-];
-
-export const 더미데이터 = [
-  "더미1",
-  "더미2",
-  "더미3",
-  "더미4",
-  "더미5",
-  "더미6",
-];
-
-export const hi = {
-  one: {
-    key: "key",
-    value: "내용내용",
-  },
-};
