--- src/components/LateNight/LateNightToday/style.ts
@@ -1,21 +1,0 @@
-import styled from "styled-components";
-
-export const LateNightHeaderContainer = styled.div`
-  display: flex;
-
-  gap: 850px;
-`;
-export const SearchBarContainer = styled.div`
-  display: flex;
-`;
-export const SelectContainer = styled.div`
-  display: flex;
-
-  gap: 10px;
-`;
-export const InfoText = styled.p`
-  font-size: 12px;
-  color: red;
-
-  margin: 4px 0px 0px 10px;
-`;
