+++ src/components/NightStudy/NightStudyModal/constant.ts
@@ -1,0 +1,8 @@
+export const NIGHTSTUDY_MODAL_ITEMS = [
+  "이름",
+  "학반",
+  "시작일",
+  "종료일",
+  "장소",
+  "휴대폰",
+];
