+++ src/components/NightStudy/NightStudyModal/style.ts
