+++ src/components/NightStudy/NightStudyToday/style.ts
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+export const NightStudyHeaderContainer = styled.div`
+  display: flex;
+
+  gap: 850px;
+`;
+export const SearchBarContainer = styled.div`
+  display: flex;
+`;
+export const SelectContainer = styled.div`
+  display: flex;
+
+  gap: 10px;
+`;
+export const InfoText = styled.p`
+  font-size: 12px;
+  color: red;
+
+  margin: 4px 0px 0px 10px;
+`;
