+++ src/constants/NightStudy/nightstudy.constant.ts
@@ -1,0 +1,11 @@
+export const NIGHTSTUDY_ALLOW_ITEMS = [
+  "이름",
+  "학반",
+  "심자 사유",
+  "시작일",
+  "종료일",
+  "장소",
+  "휴대폰",
+  "필요 이유",
+  "승인 여부",
+];
