+++ src/queries/queryKey.ts
@@ -13,8 +13,8 @@ export const QUERY_KEYS = Object.freeze({
   offbasepass: {
     getOffBasePass: (date: string) => ["/out/date", date],
   },
-  latenight: {
-    getPendingLateNight: "/nightstudy/pending",
-    getLateNightList: "/nightstudy",
+  nightstudy: {
+    getPendingNightStudy: "/nightstudy/pending",
+    getNightStudyList: "/nightstudy",
   },
 });
