--- src/repositories/LateNight/latenight.repository.ts
@@ -1,8 +1,0 @@
-import { LateNightResponse } from "../../types/LateNight/latenight.type";
-
-export interface LateNightRepository {
-  getPendingLateNight(): Promise<LateNightResponse>;
-  getLateNightList(): Promise<LateNightResponse>;
-  patchLateNightAllow(id: number): Promise<void>;
-  patchLateNightCancel(id: number): Promise<void>;
-}
