+++ src/stores/NightStudy/nightstudy.store.ts
@@ -1,0 +1,12 @@
+import { atom } from "recoil";
+import { NightStudyType } from "types/NightStudy/nightstudy.type";
+
+export const NightStudyGrade = atom<string>({
+  key: "NightStudyGrade",
+  default: "모든학년",
+});
+
+export const NightStudyDataAtom = atom<NightStudyType | null>({
+  key: "NightStudyDataAtom",
+  default: null,
+});
